+++ crates/git_cmd/src/lib.rs
@@ -231,7 +231,8 @@ impl Repo {
 
     #[instrument(skip(self))]
     pub fn checkout(&self, object: &str) -> anyhow::Result<()> {
-        self.git(&["checkout", object])?;
+        self.git(&["checkout", object])
+            .context("failed to checkout")?;
         Ok(())
     }
 
@@ -340,6 +341,16 @@ impl Repo {
             .git(&["tag", "-l", tag])
             .context("cannot determine if git tag exists")?;
         Ok(output.lines().count() >= 1)
+    }
+
+    pub fn get_branches_of_commit(&self, commit_hash: &str) -> anyhow::Result<Vec<String>> {
+        let output = self.git(&["branch", "--contains", commit_hash])?;
+        let branches = output
+            .lines()
+            .filter_map(|l| l.split_whitespace().last())
+            .map(|s| s.to_string())
+            .collect();
+        Ok(branches)
     }
 }
 
@@ -524,5 +535,15 @@ D  crates/git_cmd/CHANGELOG.md
         }
         repo.tag("v1.0.0", "test").unwrap();
         assert!(!repo.tag_exists("v2.0.0").unwrap());
+    }
+
+    #[test]
+    fn is_branch_of_commit_detected_correctly() {
+        test_logs::init();
+        let repository_dir = tempdir().unwrap();
+        let repo = Repo::init(&repository_dir);
+        let commit_hash = repo.current_commit_hash().unwrap();
+        let branches = repo.get_branches_of_commit(&commit_hash).unwrap();
+        assert_eq!(branches, vec![repo.original_branch()]);
     }
 }