+++ crates/cargo_utils/CHANGELOG.md
@@ -6,6 +6,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.1.59](https://github.com/MarcoIeni/release-plz/compare/cargo_utils-v0.1.58...cargo_utils-v0.1.59) - 2024-10-25
+
+### Other
+
+- update Cargo.toml dependencies
+
 ## [0.1.58](https://github.com/MarcoIeni/release-plz/compare/cargo_utils-v0.1.57...cargo_utils-v0.1.58) - 2024-10-05
 
 ### Other